refactor: root.component to ts
