+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/data/items.data.ts
@@ -1,0 +1,16 @@
+import { Item } from '../types/item.model';
+
+export const items: Item[] = [
+  {
+    name: 'Tomatoes',
+    amount: 2,
+    unit: 'kg',
+    price: { value: 20, currency: 'PLN' },
+  },
+  {
+    name: 'Banana',
+    amount: 5,
+    unit: 'kg',
+    price: { value: 36, currency: 'PLN' },
+  },
+];
