+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/main.ts
@@ -1,4 +1,4 @@
-import { root } from './shop/root.component.js';
+import { root } from './shop/root.component';
 
 const $app: HTMLBodyElement | null = document.querySelector('#app');
