+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/shop/root.component.ts
@@ -1,0 +1,11 @@
+import { div } from '../framework/dom-creators';
+import { cartPanel } from './cart-panel.component';
+import { hero } from './hero.component';
+import { items } from '../data/items.data';
+
+export function root() {
+  const $hero = hero({ title: 'Shopping App', subTitle: 'buy and sell' });
+  const $container = div('container');
+  $container.append($hero, cartPanel({ items }));
+  return $container;
+}
