+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/types/item.model.ts
@@ -1,0 +1,8 @@
+import { Price } from './price.type';
+
+export type Item = {
+  name: string;
+  amount: number;
+  unit: 'g' | 'kg' | 'dag' | 'ton';
+  price: Price;
+};
