+++ week-1-ts-ecosystem/w01e02-enter-the-world-of-typescript/src/types/price.type.ts
@@ -1,0 +1,4 @@
+export type Price = {
+  value: number;
+  currency: 'PLN' | 'USD' | 'EUR' | 'GBP';
+};
